requirements.txt
matplotlib==3.8.2
matplotlib-inline==0.1.6
numpy==1.26.1
pandas==2.1.2
plotly==5.18.0
snowflake-connector-python==3.5.0
streamlit==1.28.1